+++ src/domain/models.rs
@@ -1,4 +1,3 @@
 pub mod credential;
-pub mod credentials;
 pub mod issuance;
 pub mod tenants;
