+++ src/domain/models/credential.rs
@@ -206,9 +206,7 @@ impl CredentialFilter {
             return false;
         }
         if let Some(types) = &self.credential_types
-            && !types
-                .iter()
-                .any(|t| credential.credential_types.contains(t))
+            && &credential.credential_types != types
         {
             return false;
         }
