+++ src/domain/models/issuance/mod.rs
@@ -1,4 +1,5 @@
 mod consent;
+mod credentials;
 mod error;
 mod events;
 mod start;
@@ -6,6 +7,7 @@ mod task;
 mod tx_code;
 
 pub use consent::{ConsentError, ConsentRequest, ConsentResponse, NextAction};
+pub use credentials::{CredentialListQuery, CredentialListResponse, CredentialRecord};
 pub use error::{IssuanceError, IssuanceErrorCode};
 pub use events::*;
 use parking_lot::Mutex;
