+++ src/domain/service.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use crate::domain::models::issuance::IssuanceEngine;
-use crate::domain::ports::{CredentialRepo, TenantRepo};
+use crate::domain::ports::TenantRepo;
 use crate::session::SessionStore;
 
 #[derive(Clone)]
@@ -9,21 +9,14 @@ pub struct Service<S> {
     pub session: S,
     pub tenant_repo: Arc<dyn TenantRepo>,
     pub issuance_engine: IssuanceEngine,
-    pub credential_repo: Arc<dyn CredentialRepo>,
 }
 
 impl<S: SessionStore + Clone> Service<S> {
-    pub fn new<R: TenantRepo>(
-        session: S,
-        tenant_repo: R,
-        issuance_engine: IssuanceEngine,
-        credential_repo: impl CredentialRepo,
-    ) -> Self {
+    pub fn new<R: TenantRepo>(session: S, tenant_repo: R, issuance_engine: IssuanceEngine) -> Self {
         Self {
             session,
             tenant_repo: Arc::new(tenant_repo),
             issuance_engine,
-            credential_repo: Arc::new(credential_repo),
         }
     }
 }
@@ -34,10 +27,6 @@ impl<S> std::fmt::Debug for Service<S> {
             .field("session", &std::any::type_name::<S>())
             .field("tenant_repo", &std::any::type_name::<dyn TenantRepo>())
             .field("issuance_engine", &self.issuance_engine)
-            .field(
-                "credential_repo",
-                &std::any::type_name::<dyn CredentialRepo>(),
-            )
             .finish()
     }
 }
