+++ src/outbound/credential.rs
@@ -292,11 +292,20 @@ impl CredentialRepo for SqlCredentialRepo {
     }
 
     async fn list(&self, filter: CredentialFilter) -> Result<Vec<Credential>> {
+        let encoded_types = filter
+            .credential_types
+            .as_ref()
+            .map(serde_json::to_string)
+            .transpose()?;
+
         // Build SQL and collect bind values together.
         let mut builder = FilterBuilder::new(&self.driver);
 
         if let Some(ref tenant_id) = filter.tenant_id {
             builder.and("tenant_id", tenant_id.to_string());
+        }
+        if let Some(ref types) = encoded_types {
+            builder.and("credential_types", types.clone());
         }
         if let Some(ref status) = filter.status {
             builder.and("status", status.as_str().to_owned());
@@ -330,12 +339,7 @@ impl CredentialRepo for SqlCredentialRepo {
             let (id, tenant_id) = row.parse_ids()?;
             self.maybe_decrypt(&id, &mut row.raw_credential, row.payload_encrypted != 0)
                 .await?;
-            let credential = row.into_credential_with_ids(id, tenant_id)?;
-            // credential_types uses containment semantics: check in Rust rather
-            // than SQL because the stored JSON array must not be exact-matched.
-            if filter.matches(&credential) {
-                out.push(credential);
-            }
+            out.push(row.into_credential_with_ids(id, tenant_id)?);
         }
         Ok(out)
     }
