+++ src/server/handlers.rs
@@ -1,16 +1,10 @@
-mod credential;
 mod health;
 mod issuance;
 mod root;
 mod tenant;
 
 // Export handlers for use in the server
-pub use credential::{get_credential, list_credentials};
 pub use health::health_check;
-pub use issuance::authorization_callback;
-pub use issuance::get_session_events;
-pub use issuance::start_issuance;
-pub use issuance::submit_consent;
-pub use issuance::submit_transaction_code;
+pub use issuance::{authorization_callback, get_credential, get_session_events, list_credentials, start_issuance, submit_consent, submit_transaction_code};
 pub use root::home;
 pub use tenant::register_tenant;
