+++ src/server/handlers/issuance/mod.rs
@@ -1,11 +1,13 @@
 mod callback;
 mod consent;
+mod credentials;
 mod events;
 mod start;
 mod tx_code;
 
 pub use callback::authorization_callback;
 pub use consent::submit_consent;
+pub use credentials::{get_credential, list_credentials};
 pub use events::get_session_events;
 pub use start::start_issuance;
 pub use tx_code::submit_transaction_code;
